+++ go_lib/cloud-data/go.mod
@@ -1,18 +1,13 @@
-module discoverer
+module github.com/deckhouse/deckhouse/go_lib/cloud-data
 
 go 1.19
 
 require (
-	github.com/deckhouse/deckhouse/go_lib/cloud-data v0.0.0-00010101000000-000000000000
+	github.com/alecthomas/kingpin v2.2.6+incompatible
+	github.com/prometheus/client_golang v1.14.0
 	github.com/sirupsen/logrus v1.9.0
-	gopkg.in/alecthomas/kingpin.v2 v2.2.6
-)
-
-require (
-	github.com/alecthomas/kingpin v2.2.6+incompatible // indirect
-	github.com/prometheus/client_golang v1.14.0 // indirect
-	k8s.io/apimachinery v0.26.3 // indirect
-	k8s.io/client-go v0.26.3 // indirect
+	k8s.io/apimachinery v0.26.3
+	k8s.io/client-go v0.26.3
 )
 
 require (
@@ -25,7 +20,6 @@ require (
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/protobuf v1.5.2 // indirect
 	github.com/google/gofuzz v1.1.0 // indirect
-	github.com/gophercloud/gophercloud v1.2.0
 	github.com/imdario/mergo v0.3.6 // indirect
 	github.com/json-iterator/go v1.1.12 // indirect
 	github.com/matttproud/golang_protobuf_extensions v1.0.1 // indirect
@@ -51,5 +45,3 @@ require (
 	sigs.k8s.io/structured-merge-diff/v4 v4.2.3 // indirect
 	sigs.k8s.io/yaml v1.3.0 // indirect
 )
-
-replace github.com/deckhouse/deckhouse/go_lib/cloud-data => ../../../../go_lib/cloud-data/
